Add refactor
