+++ .gitignore
@@ -12,3 +12,4 @@
 openfisca_uk/
 openfisca_uk_data/
 policyengine-client/src/countries/debug.jsx
+docs/_build/
