+++ docs/_toc.yml
@@ -1,0 +1,1 @@
+root: index
