+++ docs/index.md
@@ -1,0 +1,3 @@
+# Home
+
+This is the documentation for PolicyEngine, a React client/Flask server application for interfacing with OpenFisca US and OpenFisca UK.
