--- policyengine/countries/__init__.py
@@ -1,3 +1,0 @@
-from policyengine.countries.country import PolicyEngineCountry
-from policyengine.countries.uk import UK
-from policyengine.countries.us import US
