--- policyengine/countries/uk/default_reform/__init__.py
@@ -1,1 +1,0 @@
-from policyengine.countries.uk.default_reform.reform import *
