+++ policyengine_old/countries/__init__.py
@@ -1,0 +1,3 @@
+from policyengine_old.countries.country import PolicyEngineCountry
+from policyengine_old.countries.uk import UK
+from policyengine_old.countries.us import US
