+++ policyengine_old/tests/server/test_server.py
@@ -1,0 +1,15 @@
+from policyengine_old.tests.server.initialisation import test_client
+
+
+def test_server_starts():
+    for url in [
+        "/",
+        "/us",
+        "/uk/policy",
+        "/uk/population-impact",
+        "/uk/household",
+        "/us/policy",
+        "/us/household",
+    ]:
+        response = test_client.get(url)
+        assert response.status_code == 200
